minor moves
