+++ src/presenter/presenterPoints.js
