+++ src/view/redactionForm.js
@@ -1,6 +1,6 @@
 import AbstractView from '../framework/view/abstract-view';
 import { getArrayFromType, getOfferName, getOfferPrice, CITIES } from '../mock/const.js';
-import { fullDate } from '../dateAPI.js';
+import { fullDate } from '../util/dateAPI.js';
 import { getDestById } from '../mock/mock.js';
 
 const createOfferTemplate = (offerIds, type) => getArrayFromType(type).map((offer) => {
