+++ src/view/sort.js
@@ -1,5 +1,5 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import { capitalize } from '../utils.js';
+import { capitalize } from '../util/utils.js';
 import { SortingType } from '../mock/const.js';
 
 const makeItemTemplate = (sort, status) => `
